Added download script to the dark mode script, added script to all songs
I added something to download the sheets as text format used by http://sky-music.herokuapp.com/ and sky studio (you might want to add something to make people understand this).
I then proceeded to add the script to all the already existing sheets, i made it as universal as possible so it should work with all of them, i tested some very old sheets and also new ones and all seem to be working but it still needs some tuning.

I might also add something to "preview" the song, by playing it in the background